Journal transaction boundaries
Append txn_id, txn_replica_id and txn_last to xrow_header structure.
txn_replica_id identifies replica where transaction was started and
txn_id identifies transaction id on that replica. As transaction id a lsn
of the first row in this transaction is used.
txn_last set to true if it is the last row in a transaction, so we could
commit transaction with last row or use additional NOP requests with
txn_last = true ans valid txn_id and txn_replica_id.
For replication all local changes moved to xrows array tail to form a
separate transaction (like autonomous transaction) because it is not
possible to replicate such transaction back to it's creator.

As encoding/deconding rules assumed:
 1. txn_replica_id is encoded only if it is not equal with replica id. This
 might have point because of replication trigger
 2. txn_id and txn_last are encoded only for multi-row transaction. So if we
 do not have txn_id in a xstream then this means that it is a single row
 transaction
This rules provides compatibility with previous xlog handling.

Needed for: 2798