sql: move constraint name to struct contraint_parse
Before this patch name of constraint being parsed was held as a separate
struct Token within parser context. Meanwhile, we also need to store
name of table which constraint is related to (in order to implement
ALTER TABLE ADD CONSTRAINT with several options: foreign key, unique,
primary key, etc). Hence, lets move constraint name to a separate
structure and save it alongside with the name of table.

Needed for #3097